Update Options.cpp to pass linting
